AI-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.67",
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.2",
    "langsmith>=0.7.17",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
